+++ packages/terra-list/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Added a `ariaSelectionStyle` prop that provides accessibility support for single and multi select lists
+
 ## 4.52.0 - (December 7, 2022)
 
 * Changed